finished the function of create waybill

+++ SinExWebApp20328381/ViewModels/ShipmentsListViewModel.cs
@@ -7,7 +7,7 @@ namespace SinExWebApp20328381.ViewModels
 {
     public class ShipmentsListViewModel
     {
-        public virtual int WaybillId { get; set; }
+        public virtual long WaybillId { get; set; }
         public virtual string ServiceType { get; set; }
         public virtual DateTime ShippedDate { get; set; }
         public virtual DateTime DeliveredDate { get; set; }
@@ -15,6 +15,6 @@ namespace SinExWebApp20328381.ViewModels
         public virtual int NumberOfPackages { get; set; }
         public virtual string Origin { get; set; }
         public virtual string Destination { get; set; }
-        public virtual int ShippingAccountId { get; set; }
+        public virtual long ShippingAccountId { get; set; }
     }
 }
